requirements.txt
asgiref==3.2.3
dataclasses==0.6
Django==3.0.4
django-filter==2.2.0
django-mongodb-engine-py3==0.6.0.1
django-rest-auth==0.9.5
django-rest-framework-mongoengine==3.4.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangotoolbox==1.8.0
djoser==2.0.3
mongoengine==0.19.1
numpy==1.18.1
pandas==1.0.1
psycopg2-binary==2.8.4
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
xlrd==1.2.0
